Modified Makefile_wpar to actually compile each file (except the sac executable). Also small modification to saccpt.f.
Each Makefile was only makeing the file shown in the first block/rule. I had to add the all: clean stuff so that make would try compiling each of the files. There is still one compilation error though, pertaining to sac. It might be with my sac installation, or another problem with saccpt.f.

I changed saccpt.f to allocate n(0:10) instead of n(10), to fix a compilation issue.